test: extract visual float tests atop b2df
